+++ internal/jenkins/types/signer.go
@@ -1,6 +1,10 @@
 package types
 
+import (
+	"encoding/json"
+)
+
 type Signer interface {
-	GetSignature(unsinged *InsecureUpdateJSON) (Signature, error)
-	VerifySignature(unsinged *InsecureUpdateJSON, signature Signature) error
+	GetSignature(unsinged json.Marshaler) (Signature, error)
+	VerifySignature(unsinged json.Marshaler, signature Signature) error
 }
