--- internal/json/json.go
@@ -1,9 +1,0 @@
-package json
-
-import (
-	"olympos.io/encoding/cjson"
-)
-
-func MarshalJSON[T any](ptr *T) ([]byte, error) {
-	return cjson.Marshal(ptr)
-}
